+++ src/Form/AdEntityDeleteForm.php
@@ -38,7 +38,7 @@ class AdEntityDeleteForm extends EntityConfirmFormBase {
   public function submitForm(array &$form, FormStateInterface $form_state) {
     $this->entity->delete();
 
-    drupal_set_message(
+    ​\​Drupal​\​Core​\​Messenger​\​MessengerInterface::addMessage(
       $this->t('content @type: deleted @label.',
         [
           '@type' => $this->entity->bundle(),