+++ src/Form/AdEntityForm.php
@@ -259,13 +259,13 @@ class AdEntityForm extends EntityForm {
 
     switch ($status) {
       case SAVED_NEW:
-        drupal_set_message($this->t('Created the %label Advertising entity.', [
+        ​\​Drupal​\​Core​\​Messenger​\​MessengerInterface::addMessage($this->t('Created the %label Advertising entity.', [
           '%label' => $ad_entity->label(),
         ]));
         break;
 
       default:
-        drupal_set_message($this->t('Saved the %label Advertising entity.', [
+        ​\​Drupal​\​Core​\​Messenger​\​MessengerInterface::addMessage($this->t('Saved the %label Advertising entity.', [
           '%label' => $ad_entity->label(),
         ]));
     }